pink-macro: Refactor file structure
